Extract semantic function analysis
